Use AssertJ for assertions
Improve the readability of tests. E.g.

    assertThat(actualPlugins)
        .containsExactlyInAnyOrder(
            new Plugin("workflow-scm-step", "2.4", null, null),
            new Plugin("workflow-step-api", "2.20", null, null));

is easier to read than

    List<Plugin> expectedPlugins = Arrays.asList(
        new Plugin("workflow-scm-step", "2.4", null, null),
        new Plugin("workflow-step-api", "2.20", null, null));
    assertEquals(expectedPlugins, actualPlugins);

and also more specific because it explicitly states that the order does
not matter.

Also the messages for failing tests is more helpful than the current
one. E.g. if the assertion from above fails it prints an error like

    Expecting:
      <[workflow-scm-step 2.4]>
    to contain exactly in any order:
      <[workflow-scm-step 2.4, workflow-step-api 2.20]>
    but could not find the following elements:
      <[workflow-step-api 2.20]>

so that it is immediately clear that one plugin is missing.