Handle optional dependencies correctly
The previous behavior around optional dependencies was incorrect in
some cases. If plugin A declares a required dependency on plugin C
at version N, and plugin B declares an optional dependency on plugin
C at version N+1, the resolver will completely ignore plugin B's
optional dependency and end up installing plugin C at version N. This
will break plugin B because plugin B is explicitly incompatible with
versions vefore its dependency.

To solve this, we keep optional plugins in the dependency graph until
the very end. If every dependency on a given plugin is optional, then
that plugin can be omitted from the resulting set. However, if any of
those plugins are non-optional, then that plugin must be included
in the result and the version of that plugin must be the highest
required version among all enumerated dependencies.