Fix resolution of non-optional deps of optional deps
The previous commit has an edge case where non-optional
dependencies of optional dependencies are considered
non-optional dependencies.  This is wrong, as those
dependencies should only be considered non-optional if
their parent dependency is also considered non-optional.

With this change, we do not expand dependencies of optional
dependencies until those dependencies are determined to be
non-optional.