Optional plugins missing from update center should not fail installation
When resolving plugin, we will eagerly check the update center
for plugin metadata, even when that plugin is optional. When the
plugin is missing from update center, we end up failing the
installation.

This doe snot make sense for optional plugins. These plugins are
likely never going to be installed, so it isn't a big deal that
we are missing some metadata. Instead of failing the build, we
just print a warning.

If another plugin does depend on this missing plugin in a
non-optional way, we will fail the installation at that point.