Suppress RuntimeException warnings
RuntimeExceptions in this context are often used to indicate a malformed
URL provided by the user. The program cannot do much with the malformed
URL except inform the user of the issue.

One of the suppressed warning cases handles more than malformed URL
exceptions. That case is noted with a comment in the code and with a
different justification.