Remove misleading CLI documentation
Most of the on/off cli flags are backed by `BooleanOptionHandler` which
does not allow explicit boolean values to be provided.
The exception is `--latest` which uses `ExplicitBooleanOptionHandler`.
This change removes the misleading "set to true" or "set to false" for
the cli flags that do not accept true/false (or any other values).