Skip security warnings heading if no warnings
Test highlights that there is ambiguity in the tested code because there
are two booleans to represent a single condition.  It is possible to
have hideWarning == true and showWarning == true, even though we can
only do one or the other but not both.

Another pull request or another change in this pull request will need
to resolve that issue in the most recently merged pull request.