Remove explicit snakeyaml dependency
Rely on the transitive dependency from the Jackson dataformat library.

Commit 6239603091be6af55c9459bb2baa1a272fa216a9 added the dependency in
order to assure that we were including the most recent release of the
snakeyaml jar file.  The Jackson dataformat library version that we are
using already depends on the most recent snakeyaml release, so we no
longer need this explicit dependency.

Reverts 6239603091be6af55c9459bb2baa1a272fa216a9